Add files via upload
Hello,

Please find attached a 2 new layouts. It includes a DEMUX 1x4 based on a directional coupler, with the following specifications:

Polarization: TE at 1550 nm

ΔL1 = 143

ΔL2 = 71.5

3 optical outputs: OPT2, OPT3, OPT4 (OPT in File2 )

This files does not replace the previous one and will not be included in the current report.
I will perform the analysis separately and provide an additional report afterwards.

Thank you.