Additional design (EBeam_adafridi_Antenna)
This is my additional design (Layout) for on-chip wireless communication. I hope you still have some space for the putting this in the merge file for the fabrication run. There are a lot of verification errors because of open end pins, etc. I am aware of these errors and they should not hinder the operation of these circuits or fabrication (I hope). The designs are based on the article: https://doi.org/10.1038/lsa.2017.53
P.S: My main MZI/MI based design/layout (EBeam_adafridi.gds) is unchanged, since this was passed in the feedback session and no changes were recommended.