My second layout
I would like to add a second optional plan, in addition to the first one I submitted, if space allows. Lucas mentioned this would be allowed as long as the real state on the wafer permits, If I have to choose between the two, I would rather keep this newer layout. Thanks,!-Zahra